ai-service/requirements.txt
fastapi
uvicorn
transformers
torch
spacy
scikit-learn
python-dotenv